requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.11.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
line-bot-sdk==1.16.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
protobuf==3.12.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
six==1.15.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
